refactor(oxc): rename oxlint setup RPC
